completed MVP

+++ src/pages/ViewCreator.tsx
@@ -3,35 +3,49 @@
 import { Link, useParams, useNavigate } from "react-router-dom";
 import { CreatorModel } from "../App";
 import { deleteCreator } from "../utlities/Functions";
+import { FaYoutube, FaTwitter, FaInstagram } from "react-icons/fa";
 
 
 const ViewCreator = ({creators, refresh}) => {
     const { id } = useParams();
     const navigate = useNavigate();
     const selected: CreatorModel = creators.find((crea) => crea.id == id)
+    
+    if (!creators) {
+        return <h3>NO CREATORS YET</h3>
+    }
+
     return (
         <div>
-            <h3>
-                ViewCreator
-            </h3>
-            <section>
-                {selected!.name}
-                {selected!.url}
-                {selected!.image_url}
-                {selected!.description}
-            </section>
-            <section>
-                <Link to={`/edit/${selected.id}/`}>
+            <div className="view-top">
+                <img src={selected!.image_url}/>
+                <div className="view-title-icons">
+                    <h3>
+                        {selected!.name}    
+                    </h3>
+                    <p>
+                        {selected!.description}
+                    </p>
+                    <div className="card-socialmedia-icons">
+                        {selected.youtube_url ? <FaYoutube style={{color:'white', fontSize: '40px'}}/> : null}
+                        {selected.twitter_url ? <FaTwitter style={{color:'white', fontSize: '40px'}}/> : null}
+                        {selected.instagram_url ? <FaInstagram style={{color:'white', fontSize: '40px'}}/> : null}
+                    </div>
+                </div>
+            </div>
+            <div className="bottom-buttons-container">
+                <Link className="bottom-buttons button-submit" to={`/edit/${selected.id}/`}>
                     edit
                 </Link>
-                <button onClick={()=> {
+                <button className="bottom-buttons button-submit delete" onClick={()=> {
+                console.log("called delete", id!)
                 deleteCreator(parseInt(id!));
                 refresh();
                 navigate(`/`);
-            }}>
+                }}>
                 Delete
-            </button>
-            </section>
+                </button>
+            </div>
         </div>
     );
 }
